- Change TrxId to TrxKey for consistency across services - Add RuleCode, RuleType, RuleDesc in transaction.proto - Add sample transaction with RC 51 (Insuffictient Fund) in simulator.go - Implement StoreRiskResult, write processed transaction + rule into database - Replace ksuid with hash as trxKey - Renew Readme.md
