+++ cmd/simulator/main.go
@@ -12,7 +12,7 @@ import (
 // RC 00
 // const hexPayload = `2025269C0094540C001C704C04359650400000205CC12025269C0094540CD5001C704C04359650400000205C001C019000100C40404040007C0000000000000CF0F0F0F0F0F00000000005255C0000000005255CF9F9F9F9F9F9F9F9F0F2F2F705999C360CF0F0F0F000000472613C99999999999CF0F0F5F0F3F7404040D6C1C1C9C1C4F4F2F0F0F1F0F1F9F0F0F0F1F0F0404040360C360CD5C8C3D540D5C1C4F4F20094540C0094540C000CD10000F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F040999999999999999C40404040404040404040404040404040404040404040400000000C40C3D7C140E6C9D9E2C1D34040404040404040404040404040404040999C4040400000000000000C0000000000000C0000000000000C000CD64040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040409999999C000CF0F0C1D3C5E7C9E240D7E4D5E8C140E9C1D4C1D5C9404040404040C4C5D5D7C1E2C1D94040404040C9C4F54040404040404040404040404040401010000000000CD5007C9999999C404040404040404040400094540C00926C404040404040400005F1F0F0F000000C09999C09999C09999C4040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000000000000C40404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040000000`
 // RC 51
-const hexPayload = `2025274C0163751C001C704C04359650400000205CC12025274C0163751CD5001C704C04359650400000205C001C019000100C40404040027C0000000000000CF0F0F0F0F0F00007272727272C0007272727272CF9F9F9F9F9F9F9F9F0F2F2F705999C360CF0F0F0F000000472613C99999999999C404040404040404040D6C1C1C9C1C5F3F2F0F0F1F0F1F9F0F0F0F1F0F0404040360C360CD5C8C3D540D5C1C5F3F20163751C0163751C007D400000F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F040999999999999999C40404040404040404040404040404040404040404040400000000C40C3D7C140E6C9D9E2C1D34040404040404040404040404040404040999C4040400000000000000C0000000000000C0000000000000C000CD64040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040409999999C000CF5F1C1D3C5E7C9E240D7E4D5E8C140E9C1D4C1D5C9404040404040C4C5D5D7C1E2C1D94040404040C9C4F54040404040404040404040404040401010000000000CD5007C9999999C404040404040404040400163751C01001C404040404040400005F1F0F0F000000C09999C09999C09999C4040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000000000000C40404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040`
+const hexPayload = `2025274C0163751C001C704C04359650400000205CC12025274C0163751CD5001C704C04359650400000205C001C019000100C40404040027C0000000000000CF0F0F0F0F0F00000072727272C0000072727272CF9F9F9F9F9F9F9F9F0F2F2F705999C360CF0F0F0F000000472613C99999999999C404040404040404040D6C1C1C9C1C5F3F2F0F0F1F0F1F9F0F0F0F1F0F0404040360C360CD5C8C3D540D5C1C5F3F20163751C0163751C007D400000F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F0F040999999999999999C40404040404040404040404040404040404040404040400000000C40C3D7C140E6C9D9E2C1D34040404040404040404040404040404040999C4040400000000000000C0000000000000C0000000000000C000CD64040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040409999999C000CF5F1C1D3C5E7C9E240D7E4D5E8C140E9C1D4C1D5C9404040404040C4C5D5D7C1E2C1D94040404040C9C4F54040404040404040404040404040401010000000000CD5007C9999999C404040404040404040400163751C01001C404040404040400005F1F0F0F000000C09999C09999C09999C4040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000000000000C40404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040000040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404000004040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040`
 
 // 2. Buat regex untuk menghapus semua yang bukan karakter HEX
 var re = regexp.MustCompile(`[^0-9a-fA-F]+`)
